backend/requirements.txt
Flask==2.2.2
requests==2.28.1
python-dotenv==0.21.0
werkzeug==2.2.2
flask-cors==3.0.10